Module-arize app logic
