+++ config.ru
@@ -1,3 +1,3 @@
-require "./app"
+require "./app/app.rb"
 
 run Sinatra::Application
